implemented the Node class and Operation enum to be used in ExpressionTree, also covered the above two with unit tests

+++ src/main/java/com/epam/training/expression/Expression.java
@@ -2,7 +2,7 @@ package com.epam.training.expression;
 
 public class Expression implements ExpressionInterface {
 
-    private Float value;
+    private Double value;
     private ExpressionTree expTree;
 
     @Override
